64 3rd-June-TwoPointers-RAG/fs_rag/server/requirements.txt
flask
flask-cors
PyPDF2
langchain
faiss-cpu
huggingface_hub
requests
pyjwt
pymongo